wasip1: add `//go:wasmexport` support
This adds support for the `//go:wasmexport` pragma as proposed here:
https://github.com/golang/go/issues/65199

It is currently implemented only for wasip1, but it is certainly
possible to extend it to other targets like GOOS=js, wasm-unknown, or
wasip2. It is also currently limited to -buildmode=c-shared, this is a
limitation that could easily be lifted in the future.