wasm/js: use standard library syscall package
This switches -target=wasm (browser wasm) over from our own syscall
package to the one used in the Go standard library.

While this doesn't remove any code (so we can't simplify anything), the
idea is that this improves compatibility with existing code a bit more.
So It's a similar reasoning as for
https://github.com/tinygo-org/tinygo/pull/4417.